Okay... forget about user_id
User ID isn't working out because slash commands don't send a USER ID 
but a user name instead. And user name's don't have to be unique which 
means figuring out who X is will be difficult. So remove all traces of 
user ID and just take a name from the "ask" command and use that 
everywhere. (Also map "ask" message timestamps with the user name to 
maintain state for when the interaction comes in)